Changed back to sql 5.7 now working properly. Password was changed in properties file Tables tools, backgrounds, and languages need to be filled out (DB issue)

+++ src/java/midknight/munch/dtable/model/Tool.java
@@ -28,7 +28,7 @@ public class Tool implements Serializable{
     @Column(name="cost")
     private int cost;
     @Column(name="weight")
-    private int weight;
+    private float weight;
     @Column(name="tool_type_name")
     private String toolTypeName;
     @Column(name="description")
@@ -66,7 +66,7 @@ public class Tool implements Serializable{
         this.cost = cost;
     }
 
-    public int getWeight() {
+    public float getWeight() {
         return weight;
     }
